--- a/tests/questionpy_server/worker/impl/test_thread.py
+++ b/tests/questionpy_server/worker/impl/test_thread.py
@@ -7,20 +7,15 @@ from unittest.mock import patch
 
 import pytest
 
-from questionpy_common.constants import MiB
 from questionpy_server import WorkerPool
 from questionpy_server.worker.impl.thread import ThreadWorker
 from tests.conftest import PACKAGE
 
 
-@pytest.fixture
-def pool() -> WorkerPool:
-    return WorkerPool(1, 512 * MiB, worker_type=ThreadWorker)
-
-
-async def test_should_ignore_limits(pool: WorkerPool) -> None:
+@pytest.mark.parametrize("worker_pool", [ThreadWorker], indirect=True)
+async def test_should_ignore_limits(worker_pool: WorkerPool) -> None:
     with patch.object(resource, "setrlimit") as mock:
-        async with pool.get_worker(PACKAGE, 1, 1):
+        async with worker_pool.get_worker(PACKAGE, 1, 1):
             pass
 
         mock.assert_not_called()