Closes #143.

Da der `ThreadWorker` die `WorkerResourceLimits` ignoriert, ist es mit der aktuellen Implementierung theoretisch möglich "unbegrenzt" viele `ThreadWorker` zu erstellen. Da der Worker nur zum Debuggen verwendet wird, habe ich keine extra Logik eingebaut, die die Anzahl der `ThreadWorker` beschränkt. 

Performance-Beispiel: Beim Aufruf von `/packages/{hash}/question` habe ich auf meiner Maschine folgende Verbesserung der Antwortzeiten erzielt:
  * `ProcessWorker` ca. 20x (500ms -> 25ms)
  * `ThreadWorker` ca.     4.5x (90ms -> 20ms)